Приступаю к закруглению углов убложек
