+++ app/src/main/java/com/practicum/playlistmaker/SearchActivity.kt
@@ -9,6 +9,8 @@ import android.widget.EditText
 import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 
 class SearchActivity : AppCompatActivity() {
 
@@ -52,6 +54,39 @@ class SearchActivity : AppCompatActivity() {
             searchText = savedInstanceState.getString(SEARCH_TEXT_KEY, "")
             searchBox.setText(searchText)
         }
+
+        // Инициализация RecyclerView и TrackAdapter
+        val recycler = findViewById<RecyclerView>(R.id.trackList)
+        val tracks = arrayListOf(
+            Track(
+                trackName = "Smells Like Teen Spirit",
+                artistNameAndTime = "Nirvana · 5:01",
+                artworkUrl100 = "https://is5-ssl.mzstatic.com/image/thumb/Music115/v4/7b/58/c2/7b58c21a-2b51-2bb2-e59a-9bb9b96ad8c3/00602567924166.rgb.jpg/100x100bb.jpg"
+            ),
+            Track(
+                trackName = "Billie Jean",
+                artistNameAndTime = "Michael Jackson · 4:35",
+                artworkUrl100 = "https://is5-ssl.mzstatic.com/image/thumb/Music125/v4/3d/9d/38/3d9d3811-71f0-3a0e-1ada-3004e56ff852/827969428726.jpg/100x100bb.jpg"
+            ),
+            Track(
+                trackName = "Stayin' Alive",
+                artistNameAndTime = "Bee Gees · 4:10",
+                artworkUrl100 = "https://is4-ssl.mzstatic.com/image/thumb/Music115/v4/1f/80/1f/1f801fc1-8c0f-ea3e-d3e5-387c6619619e/16UMGIM86640.rgb.jpg/100x100bb.jpg"
+            ),
+            Track(
+                trackName = "Whole Lotta Love",
+                artistNameAndTime = "Led Zeppelin · 5:33",
+                artworkUrl100 = "https://is4-ssl.mzstatic.com/image/thumb/Music115/v4/1f/80/1f/1f801fc1-8c0f-ea3e-d3e5-387c6619619e/16UMGIM86640.rgb.jpg/100x100bb.jpg"
+            ),
+            Track(
+                trackName = "Sweet Child O'Mine",
+                artistNameAndTime = "Guns N' Roses · 5:03",
+                artworkUrl100 = "https://is5-ssl.mzstatic.com/image/thumb/Music125/v4/a0/4d/c4/a04dc484-03cc-02aa-fa82-5334fcb4bc16/18UMGIM24878.rgb.jpg/100x100bb.jpg"
+            )
+        )
+
+        recycler.layoutManager = LinearLayoutManager(this)
+        recycler.adapter = TrackAdapter(tracks)
     }
 
     override fun onSaveInstanceState(outState: Bundle) {
@@ -74,4 +109,3 @@ class SearchActivity : AppCompatActivity() {
         private const val SEARCH_TEXT_KEY = "SEARCH_TEXT"
     }
 }
-
