+++ app/src/main/res/layout/item_track_card.xml
@@ -11,7 +11,7 @@
         style="@style/cover_style"
         android:contentDescription="@string/album_cover"
         android:src="@drawable/text_cover"
-        android:scaleType="centerCrop"/>
+        android:scaleType="centerCrop" />
 
     <LinearLayout
         android:layout_width="0dp"
