+++ app/src/main/res/values/styles.xml
@@ -44,6 +44,7 @@
         <item name="android:layout_height">wrap_content</item>
         <item name="android:padding">@dimen/activity_settings_padding_20dp</item>
 
+
     </style>
 
     <style name="settingsSwitch">
@@ -96,6 +97,7 @@
         <item name="android:layout_height">45dp</item>
         <item name="android:layout_width">45dp</item>
         <item name="android:layout_margin">8dp</item>
+
     </style>
 
     <style name="track_name_style">
